Set a code chunk

You can use these two functions to set code chunks within a reactive inside the server function of a teal module. You can use %chunk>% if you want to simply set the call into the chunk. You can use <%chunk>% if you also want to assign the left hand side value to the result of this expression

Examples

session <- new.env()
session$userData <- new.env()
session$ns <- function(x) {
  if (length(x) == 0) {
    "id"
  } else {
    paste0("id", x, sep = "-")
  }
}
session$userData[[session$ns(character(0))]]$chunks <- chunks$new()

x <- 5
session$userData[[session$ns(character(0))]]$chunks$reset()
myval %<chunk% sum(x, 3)

mysum %<chunk% sum(x, 5)

chunks_eval(chunks = session$userData[[session$ns(character(0))]]$chunks)
#> [1] 10
#>[1] 10

cat(
  paste(
    session$userData[[session$ns(character(0))]]$chunks$get_rcode(),
    collapse = "\n"
  )
)
#> sum(x, 3)
#> sum(x, 5)
# x + 3
# sum(x, 5)